feat: snakemake data pipeline framework
This commit introduces a snakemake framework and a CLI script that exposes the functionality of the IceFloeTracker.jl module. This work assumes that we will be processing batches of files on the same node. We may find later that we want to rework out pipeline so that single images are processed on nodes, and those images are broken down into many tiles that are then worked on in parallel.